+++ constants/DSnakeLab.py
@@ -8,8 +8,9 @@ class DSnakeLab:
     DB_NAME: Final[str] = "snakelab"
     DB_PORT: Final[int] = 3306
     DB_USER: Final[str] = "snakelab"
+    EVENTS_PORT: Final[int] = 41972
     PORT: Final[int] = 41970
     TELEMETRY_FRAME_RATE: Final[float] = 15.0
     TELEMETRY_PORT: Final[int] = 41971
     SERVER_LOG_FILE: Final[str] = "/opt/snake-lab/logs/server.log"
-    VERSION: Final[str] = "0.9.2"
+    VERSION: Final[str] = "0.9.3"